Timeout in start_new_kernel if kernel fails to respond
Closes jupyter/nbconvert#217

start_new_kernel has a startup_timeout parameter, but it wasn't being
used anywhere. I pass it through to KernelClient.wait_for_ready(), and
raise RuntimeError if it times out. (Maybe should raise something else?
TimeoutError? But this is for OS level operations with errno ETIMEDOUT -
not sure if we should use it for arbitrary timeouts).

It also checks whether the kernel is still alive, so if the kernel has
already died, we don't need to wait 60 seconds before giving up. It will
still wait one second trying to get the message from the kernel.