Avoid passing down unexpected kwargs
Squashes some warnings from recent traitlets changes.

There's still another issue with KernelClient - it's passed parameters
from the connection info which it doesn't use (key and
signature_scheme), because it's also passed a reference to a Session
object which already has those things. I'm not sure of the neatest way
to resolve that.