Allow configuration of kernel shutdown waittime
This adds a 'shutdown_wait_time' trait to KernelManager, giving the
number of seconds to wait after a kernel shutdown was requested before
forcibly killing it. It also ups the defaults from 1 to 5 seconds.

Some kernels might want to do things at the end of the session, such as
committing recorded usage, cleanup connections, etc, and having them die
is unfortunate.