Add a `metadata` field to kernelspecs.
Currently, the only way for a kernel to offer additional information about its
capabilities is to encode information into the kernel name or display name;
this information can be useful to clients, especially for things like
filtering a list of kernels.

This adds support for a new kernelspec dict field, `metadata`. This allows
kernels to add additional information, which clients can then consume as
needed.