Backport PR #294: Additional to the actual signal, send a message on the control port
In Erlang I have only on POSIX and only in the newest version some control over the normal runtime signals, in particular SIGINT. However, I can easily pick up and process the respective message on the control socket to interrupt execution.

Can this functionality be added? If yes, I'd continue and document the functionality and add tests. Maybe it would also be good to have the option of not sending an actual signal (via `os.kill`) at all.

Signed-off-by: Thomas Kluyver <thomas@kluyver.me.uk>