Backport PR #352: tornado 5 fixes in ThreadedClient
The main issue was failing to schedule `ioloop.stop` in the ioloop thread, which is required when tornado is running on asyncio in order to wake the thread. The result of not doing this is hanging forever when trying to exit, e.g. in QtConsole.

There is further cleanup of threadsafety issues with respect to asyncio and tornado objects.

closes https://github.com/jupyter/qtconsole/issues/275

cc  ccordoba12

Signed-off-by: Min RK <benjaminrk@gmail.com>