Wrap setting of kernel_id with method that can then be overridden in subclasses.
A recent requirement for Jupyter Enterprise Gateway is for clients to be able to
specify the kernel_id for new kernels.  Although `jupyter_client.start_kernel()` will
honor client-provided kernel_ids, Notebook's override of `start_kernel()` changes
the semantics of a non-null kernel_id in the argument list to mean an existing
(persisted) kernel should be _started_.  As a result, applications that derive
from the kernel management infrastructure beyond Notebook cannot influence the
derivation of the kernel's id via the existing argument list behavior.

By introducing the `determine_kernel_id()` method, subclasses are able to derive
the kernel's id however they wish.

With the ability to know the kernel's id prior to its invocation, a number of
things can be done that wouldn't be possible otherwise.  For example, this
provides the ability to setup a shared filesystem location possibly pre-populated
with data relative to what the request (i.e., kernel) is going to need.