Allow multi kernel manager to handle asynchronous start kernel
This is a work in progress to allow asynchronous start of kernels,
typically kernel that will be submitted via job queue and for which you
do not know (yet), the host on which they will be running.

This is work in progress, but I'm suspecting we want to push for async
methods in a lot of place, and there is imho no hard in making them
async as they don't _have to_ yield.

We can't (AFAICT) change start_kernel to be async as this would be a
change of API for our consumers.