Avoid trying to set the sticky bit of a regular file
As it is , the paths list starts with the saved .json file, subsequently setting the sticky bit of its directory. There is no need to set the sticky bit of a regular file, it only generates otherwise a harmless message. No modern UNIX(-like) OS supports sticky bit functionality now, as far as I am aware.