Add some nest_asyncio insurance
This does not guarantee a convenient patch time, but has a much higher chance of hitting the right timing. It also only applies the patch if a sync class is actually instantiated, and has minimal overhead for async classes.

A rather blunt instrument, but also rather hard to solve generically without putting weird restrictions on consumers of jupyter-client.