Minor Fixes.
I'm investigating an issue where Jupyter-Console does not and it seem
there are mixup between sync and async KernelManager and other in a few
places.

Basically both Async and Blocking KM are expected by the type system to
work as KM independently, but obviously some have Async methods and
other have Sync methods, which need to be awaited.

When touching traitlets config files it's easy to change the default
class from Sync to async, and then you get weird error messages.

So dropping a few asserts here and fixign a few types while
investigating.

I think the Sync and Async should be more distinct now that we are
mostly async everywhere to avoid those confusions.